removed common in server, moved code, removed more dead code
