+++ conans/server/rest/controller/v2/ping.py
