+++ conans/server/service/v2/service_v2.py
@@ -4,13 +4,13 @@ import os
 from bottle import FileUpload, static_file
 
 from conans.errors import RecipeNotFoundException, PackageNotFoundException, NotFoundException
-from conans.server.service.common.common import CommonService
+from conans.model.package_ref import PkgReference
 from conans.server.service.mime import get_mime_type
 from conans.server.store.server_store import ServerStore
 from conans.util.files import mkdir
 
 
-class ConanServiceV2(CommonService):
+class ConanServiceV2:
 
     def __init__(self, authorizer, server_store):
         assert(isinstance(server_store, ServerStore))
@@ -119,3 +119,26 @@ class ConanServiceV2(CommonService):
         if not os.path.exists(os.path.dirname(path)):
             mkdir(os.path.dirname(path))
         file_saver.save(os.path.dirname(path))
+
+    # REMOVE
+    def remove_recipe(self, ref, auth_user):
+        self._authorizer.check_delete_conan(auth_user, ref)
+        self._server_store.remove_recipe(ref)
+
+    def remove_package(self, pref, auth_user):
+        self._authorizer.check_delete_package(auth_user, pref)
+
+        for rrev in self._server_store.get_recipe_revisions_references(pref.ref):
+            new_ref = copy.copy(pref.ref)
+            new_ref.revision = rrev.revision
+            # FIXME: Just assign rrev when introduce RecipeReference
+            new_pref = PkgReference(new_ref, pref.package_id, pref.revision)
+            for _pref in self._server_store.get_package_revisions_references(new_pref):
+                self._server_store.remove_package(_pref)
+
+    def remove_all_packages(self, ref, auth_user):
+        self._authorizer.check_delete_conan(auth_user, ref)
+        for rrev in self._server_store.get_recipe_revisions_references(ref):
+            tmp = copy.copy(ref)
+            tmp.revision = rrev.revision
+            self._server_store.remove_all_packages(tmp)
